requirements.txt
# Django
Django
django-extensions

# Django REST Framework
djangorestframework
django-filter

# JWT
djangorestframework_simplejwt

# TESTING
flake8